+++ packages/rpc/src/adaptors/message-port.ts
@@ -1,6 +1,7 @@
 import type {MessageEndpoint} from '../types';
 
 export function fromMessagePort(messagePort: MessagePort): MessageEndpoint {
+  messagePort.start();
   return {
     postMessage: (...args: [any, Transferable[]]) =>
       messagePort.postMessage(...args),